backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
authlib
python-dotenv
jinja2
firebase-admin
gunicorn
itsdangerous
google-auth
google-auth-oauthlib
openai
aiosqlite
sqlalchemy
httpx
greenlet
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
